Move database to MongoDB atlas
